+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aave/react
 
+## 4.0.0-next.13
+
+### Patch Changes
+
+- 93adceb: **fix:** `useTokenSwap` and `useCancelSwap` hooks.
+- Updated dependencies [93adceb]
+  - @aave/graphql@1.0.0-next.12
+  - @aave/client@4.0.0-next.13
+
 ## 4.0.0-next.12
 
 ### Patch Changes